# postman/document-processing.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Document Extraction API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/config/document-types/{document_type}/profiles/{profile_name}": {
      "post": {
        "tags": [
          "Configuration API v1"
        ],
        "summary": "Add a document configuration profile",
        "description": "Adds one uniquely named profile to a profile-based document type. Legacy document types are rejected. Authenticated requests require the config:write scope.",
        "operationId": "add_profile_api_v1_config_document_types__document_type__profiles__profile_name__post",
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "document_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Profile-based document-type configuration key.",
              "title": "Document Type"
            },
            "description": "Profile-based document-type configuration key."
          },
          {
            "name": "profile_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "pattern": "^[a-z][a-z0-9_]*$",
              "description": "New profile name matching the profile-name contract.",
              "title": "Profile Name"
            },
            "description": "New profile name matching the profile-name contract."
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional strong configuration ETag. A stale value rejects the mutation.",
              "title": "If-Match"
            },
            "description": "Optional strong configuration ETag. A stale value rejects the mutation."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddProfileRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Profile creation result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Configuration API v1"
        ],
        "summary": "Delete a document configuration profile",
        "description": "Deletes one existing non-default profile from a profile-based document type. Authenticated requests require the config:write scope.",
        "operationId": "delete_profile_api_v1_config_document_types__document_type__profiles__profile_name__delete",
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "document_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Profile-based document-type configuration key.",
              "title": "Document Type"
            },
            "description": "Profile-based document-type configuration key."
          },
          {
            "name": "profile_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "pattern": "^[a-z][a-z0-9_]*$",
              "description": "Existing non-default profile name to delete.",
              "title": "Profile Name"
            },
            "description": "Existing non-default profile name to delete."
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional strong configuration ETag. A stale value rejects the mutation.",
              "title": "If-Match"
            },
            "description": "Optional strong configuration ETag. A stale value rejects the mutation."
          }
        ],
        "responses": {
          "200": {
            "description": "Profile deletion result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/config/document-types/{document_type}/collections/{collection_name}/fields": {
      "post": {
        "tags": [
          "Configuration API v1"
        ],
        "summary": "Add a document collection field",
        "description": "Adds one uniquely named extraction field to an existing document collection. Authenticated requests require the config:write scope.",
        "operationId": "add_collection_field_api_v1_config_document_types__document_type__collections__collection_name__fields_post",
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "document_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Document-type configuration key.",
              "title": "Document Type"
            },
            "description": "Document-type configuration key."
          },
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Existing collection name.",
              "title": "Collection Name"
            },
            "description": "Existing collection name."
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional strong configuration ETag. A stale value rejects the mutation.",
              "title": "If-Match"
            },
            "description": "Optional strong configuration ETag. A stale value rejects the mutation."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddFieldRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Collection field creation result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/config/document-types/{document_type}/profiles/{profile_name}/collections/{collection_name}/fields": {
      "post": {
        "tags": [
          "Configuration API v1"
        ],
        "summary": "Add a profile collection field",
        "description": "Adds one uniquely named extraction field to an existing profile collection. Authenticated requests require the config:write scope.",
        "operationId": "add_profile_collection_field_api_v1_config_document_types__document_type__profiles__profile_name__collections__collection_name__fields_post",
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "document_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Profile-based document-type key.",
              "title": "Document Type"
            },
            "description": "Profile-based document-type key."
          },
          {
            "name": "profile_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Existing profile name.",
              "title": "Profile Name"
            },
            "description": "Existing profile name."
          },
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Existing profile collection name.",
              "title": "Collection Name"
            },
            "description": "Existing profile collection name."
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional strong configuration ETag. A stale value rejects the mutation.",
              "title": "If-Match"
            },
            "description": "Optional strong configuration ETag. A stale value rejects the mutation."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddFieldRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Profile collection field creation result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddFieldRequest": {
        "properties": {
          "field": {
            "$ref": "#/components/schemas/DocumentFieldModel",
            "description": "Extraction field configuration to add."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "field"
        ],
        "title": "AddFieldRequest"
      },
      "AddProfileRequest": {
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/DocumentProfileModel",
            "description": "Complete profile configuration to add."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "profile"
        ],
        "title": "AddProfileRequest"
      },
...
      "DocumentProfileModel": {
        "properties": {
          "matching": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SupplierMatchingModel"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional configuration-driven supplier matching rules."
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/DocumentFieldModel"
            },
            "type": "array",
            "title": "Fields",
            "description": "Profile-specific extraction fields."
          },
          "collections": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/DocumentCollectionModel"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collections",
            "description": "Profile-specific collections keyed by collection name."
          },
          "summary_validations": {
            "items": {
              "$ref": "#/components/schemas/CollectionSummaryValidationModel"
            },
            "type": "array",
            "title": "Summary Validations",
            "description": "Cross-field collection summary validations."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DocumentProfileModel"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OperationResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "ok",
            "title": "Status",
            "description": "Successful configuration operation status.",
            "examples": [
              "ok"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable result of the configuration operation.",
            "examples": [
              "Document type 'synthetic_contract' created"
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "OperationResponse"
      },
...
      "SupplierMatchingModel": {
        "properties": {
          "any_of": {
            "items": {
              "$ref": "#/components/schemas/SupplierMatchRuleModel"
            },
            "type": "array",
            "minItems": 1,
            "title": "Any Of",
            "description": "Supplier match rules where any matching pattern selects the profile."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "any_of"
        ],
        "title": "SupplierMatchingModel"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "description": "Service API key",
        "in": "header",
        "name": "X-API-Key"
      },
      "BearerAuth": {
        "type": "http",
        "description": "OIDC access token",
        "scheme": "bearer"
      }
    }
  }
}